src/image_recognition/requirements.txt
Django==2.2.5
django-allauth==0.40.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-extensions==2.2.5
django-notifications-hq==1.5.0
django-lazysignup
django-widget-tweaks

celery==4.3.0
django-celery-beat==1.5.0
redis

imutils==0.5.2
beautifulsoup4
django-picklefield

# matplotlib==3.1.1
# nltk==3.4.5
# pandas==0.25.3
# scikit-learn==0.21.3
# seaborn==0.9.0
# tensorflow==1.15.0
# psycopg2
# flask
# flask-restful

